check delete item

+++ src/Component/Setting/LoaiHoSo.js
@@ -2,6 +2,8 @@ import React from 'react';
 import 'antd/dist/antd.css';
 import { Form, Menu,Input, Button, notification, message, Modal, Table, Icon, Select, Row} from 'antd';
 import loaihosoService from '../../Service/loaihoso.service';
+import hosoService from '../../Service/hosotailieu.service';
+import tailieuService from'../../Service/tailieu.service';
 import ExportExel from '../Common/Export/ExportExel';
 import _ from 'lodash';
 const { Option } = Select;
@@ -18,6 +20,8 @@ export default class LoaiHoSo extends React.Component {
             dataLoaiHoSos: [],
             count: 1,
         };
+        this.hosoService = new hosoService();
+        this.tailieuService = new tailieuService();
         this.loaihosoService = new loaihosoService();
         this.handChange = this.handChange.bind(this);
         this.removeItem = this.removeItem.bind(this);
@@ -100,6 +104,14 @@ export default class LoaiHoSo extends React.Component {
         }
 
     }
+
+    canDelete = (record)=>{
+        if((_.indexOf(this.state.dataHoSos,record.IdLoaiHoSo) !==-1)||((_.indexOf(this.state.dataTLs,record.IdLoaiHoSo) !==-1))){   
+            return false;
+        }
+        return true;
+    }
+
     removeItem(record) {
         var _this = this;
         Modal.confirm({
@@ -108,6 +120,16 @@ export default class LoaiHoSo extends React.Component {
             okType: 'danger',
             cancelText: 'No',
             onOk() {
+                if(!_this.canDelete(record)){
+                    notification.error(
+                        {
+                            message: "Có lỗi xảy ra !",
+                            defaultValue: "topRight",
+                            description:"Dữ liệu đang được liên kết."
+                        }
+                    )
+                    return;
+                }
                 _this.loaihosoService.deleteItem(record.IdLoaiHoSo)
                     .then(function () {
                         notification.success({
@@ -156,6 +178,16 @@ export default class LoaiHoSo extends React.Component {
                     dataLoaiHoSos: data.data,
                 })
             })
+            _this.hosoService.getItems("").then(function(data){
+                _this.setState({
+                    dataHoSos:_.map(data.data,"LoaiHoSoId")
+                })
+            })
+            _this.tailieuService.getItems("").then(function(data){
+                _this.setState({
+                    dataTLs:_.map(data.data,"LoaiTaiLieuId")
+                })
+            })    
     }
     isEditting(record) {
         var _this = this;
